split reference classes into separate files for clarity
