+++ packages/algo-ts-testing/src/test-execution-context.ts
@@ -3,7 +3,9 @@ import algosdk from 'algosdk'
 import { captureMethodConfig } from './abi-metadata'
 import { DecodedLogs, decodeLogs, LogDecoding } from './decode-logs'
 import * as ops from './impl'
-import { AccountCls, ApplicationCls, AssetCls } from './reference'
+import { AccountCls } from './impl/account'
+import { ApplicationCls } from './impl/application'
+import { AssetCls } from './impl/asset'
 import { ContractContext } from './subcontexts/contract-context'
 import { LedgerContext } from './subcontexts/ledger-context'
 import { TransactionContext } from './subcontexts/transaction-context'
